Fixes after initial review
Removed all unneeded ‘self.’ code, did not know I could test using the
compiler, good call.
Fixed the constraints on the meme editor so the picture was laid out
correctly with text in bot landscape and portrait mode
Changed the detail view to a modal window and used the meme image
instead of text boxes etc.
Cancel button returns to the meme screen correctly. I still clear the
screen before I leave though so nothing is lingering. Is this needed?
Modified the amount the screen moves up when the keyboard is in play to
allow for text to be seen in landscape mode